Migrate Loader component to TypeScript
